✨ replace demo user
Ref: #137

+++ api/prisma/seed.ts
@@ -1,29 +1,26 @@
-import { CompetencyFramework, PrismaClient } from '@prisma/client';
+import { CompetencyFramework, User, PrismaClient } from '@prisma/client';
 
 import { CompetencyFixture, competencyFixtures } from './fixtures/competencies';
 
 const prisma = new PrismaClient();
 
-const sample = <T>(array: T[]): T =>
-  array[Math.floor(Math.random() * array.length)] as T;
-
 async function createCompetency(
   framework: CompetencyFramework,
   fixture: CompetencyFixture,
-  peopleIds: string[],
+  user: User,
   nesting?: { root: string; parent: string },
 ): Promise<void> {
   const competency = await prisma.competency.create({
     data: {
       competencyFrameworkId: framework.id,
-      createdById: sample(peopleIds),
-      updatedById: sample(peopleIds),
+      createdById: user.id,
+      updatedById: user.id,
       parentCompetencyId: nesting?.parent,
       translations: { create: { languageCode: 'EN', title: fixture.title } },
     },
   });
   if (Array.isArray(fixture.children)) {
-    await createCompetencies(framework, fixture.children, peopleIds, {
+    await createCompetencies(framework, fixture.children, user, {
       parent: competency.id,
       root: nesting?.root ?? competency.id,
     });
@@ -33,47 +30,44 @@ async function createCompetency(
 async function createCompetencies(
   framework: CompetencyFramework,
   fixtures: CompetencyFixture[],
-  peopleIds: string[],
+  user: User,
   nesting?: { root: string; parent: string },
 ): Promise<void> {
   for (const fixture of fixtures) {
-    await createCompetency(framework, fixture, peopleIds, nesting);
+    await createCompetency(framework, fixture, user, nesting);
   }
 }
 
 async function main(): Promise<void> {
-  await prisma.person.deleteMany({});
-  await prisma.person.createMany({
-    data: [
-      { firstName: 'Sam', lastName: 'Bossant', role: 'TEACHER' },
-      { firstName: 'Peter', lastName: 'Buytaert', role: 'TEACHER' },
-      { firstName: 'Yannick', lastName: 'Gils', role: 'TEACHER' },
-      { firstName: 'Dieter', lastName: 'Luypaert ', role: 'TEACHER' },
-      { firstName: 'Stijn', lastName: 'Van Vlierberghe', role: 'TEACHER' },
-      { firstName: 'Alexander', lastName: 'Vandoren', role: 'TEACHER' },
-    ],
+  const { SEED_USER_ID, SEED_USER_NAME } = process.env;
+  if (SEED_USER_ID == null || SEED_USER_NAME == null) {
+    throw new Error('Configure your seed user. See README');
+  }
+  await prisma.user.deleteMany({
+    where: { id: SEED_USER_ID },
   });
-  const peopleWithIds = await prisma.person.findMany({ select: { id: true } });
-  const peopleIds = peopleWithIds.map(({ id }) => id);
+  const user = await prisma.user.create({
+    data: { id: SEED_USER_ID, name: SEED_USER_NAME },
+  });
 
   await prisma.competencyFramework.deleteMany({ where: {} });
   const competencyFramework = await prisma.competencyFramework.create({
     data: {
-      createdById: sample(peopleIds),
-      updatedById: sample(peopleIds),
+      createdById: user.id,
+      updatedById: user.id,
       translations: { create: { languageCode: 'EN', title: 'CanMEDS 2015' } },
     },
   });
 
   await prisma.competency.deleteMany({ where: {} });
-  await createCompetencies(competencyFramework, competencyFixtures, peopleIds);
+  await createCompetencies(competencyFramework, competencyFixtures, user);
 
   await prisma.education.deleteMany({ where: {} });
   for (const educationTitle of ['Informatics', 'Chemistry', 'Medicine']) {
     await prisma.education.create({
       data: {
-        createdById: sample(peopleIds),
-        updatedById: sample(peopleIds),
+        createdById: user.id,
+        updatedById: user.id,
         translations: { create: { languageCode: 'EN', title: educationTitle } },
       },
     });
